Enabled MPEG layer 1 and 2 on start of playback
Added the VS1053_MODE_SM_LAYER12 flag to VS1053_REG_MODE in the startPlayingFile( ) function. This makes it possible to play MP1 or MP2 files.

Adding this flag is enough to let the VS1053 play .MPP files from the DCC-Studio program without further modification of media files or code. The files are encoded in PASC, the predecessor to MPEG-1 layer 1 which was used by Philips Digital Compact Cassette.

Playing files in other formats still works as before with the flag added. The flag is only there so that system designers who didn't want to pay for a license for those formats could let a player deny playing them. This is all moot now that the patents on MPEG have expired.